% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_dtm_dataframes}
\alias{get_dtm_dataframes}
\title{Transform normalized vectors to dataframes}
\usage{
get_dtm_dataframes(x)
}
\arguments{
\item{x}{a data frame the first column is for the terms, the second for the value(relevance)}
}
\description{
Normalize each keyword vector and transform it to data frame
}
