% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heatmap_list}
\alias{heatmap_list}
\title{Melted similarity matrix}
\usage{
heatmap_list(similarity_matrix, x_axis_lodf, y_axis_lodf, report = TRUE)
}
\arguments{
\item{similarity_matrix}{an unnamed similarity matrix (usually created with \code{\link{heatmap_matrix}})}

\item{x_axis_lodf}{a list of data frames. The first column is for the terms, the second for the value(relevance)}

\item{y_axis_lodf}{a list of data frames. The first column is for the terms, the second for the value(relevance)}

\item{report}{default to TRUE. whether to save the similarity matrix as .csv file}
}
\description{
Similarity matrix is transformed to a list, ordered from the highest similarity score. Top keywords are added for comparisson.
}
