% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heatmap_matrix}
\alias{heatmap_matrix}
\title{Get the similarity matrix}
\usage{
heatmap_matrix(x_axis_lodf, y_axis_lodf, x_label = "X", y_label = "Y",
  report = TRUE)
}
\arguments{
\item{x_axis_lodf}{a list of data frames. The first column is for the terms, the second for the value(relevance)}

\item{y_axis_lodf}{a list of data frames. The first column is for the terms, the second for the value(relevance)}

\item{x_label}{default to "X". a name for the values in the X axis}

\item{y_label}{default to "Y". a name for the values in the Y axis}

\item{report}{default to TRUE. whether to save the similarity matrix as .csv file}
}
\description{
A similarity matrix based on cosine similarity
}
