% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heatmap_viz}
\alias{heatmap_viz}
\title{Visualize the heatmap}
\usage{
heatmap_viz(edges, x_label = "X", y_label = "Y")
}
\arguments{
\item{edges}{a melted similarity matrix created with \code{\link{heatmap_list}}}

\item{x_label}{default to "X". a name for the values in the X axis}

\item{y_label}{default to "Y". a name for the values in the Y axis}
}
\description{
Plotly implementation for the similarity matrix
}
